# Speed up writing of FITS tables with string columns