--- a/astropy/io/fits/fitsrec.py
+++ b/astropy/io/fits/fitsrec.py
@@ -1291,22 +1291,37 @@ def _rstrip_inplace(array):
     calculation.
     """
 
-    # The following implementation uses a 1D view of the buffer (via ravel())
-    # converts it to a 2D array of unsigned integers. Trailing spaces (which are
-    # represented as 32 are then converted to null characters, represented as
-    # zeros). The loop over j is to avoid creating too large temporary arrays in
-    # memory.
+    # The following implementation convert the string to unsigned integers of
+    # the right length. Trailing spaces (which are represented as 32) are then
+    # converted to null characters (represented as zeros). To avoid creating
+    # large temporary mask arrays, we loop over chunks (attempting to do that
+    # on a 1-D version of the array; large memomry may still be needed in the
+    # unlikely case that a string array has small first dimension and cannot
+    # be represented as a contiguous 1-D array in memory).
 
     dt = array.dtype
 
     if dt.kind not in 'SU':
         raise TypeError("This function can only be used on string arrays")
-
-    dt_int = "{0}{1}u{2}".format(dt.itemsize // dt.alignment, dt.byteorder, dt.alignment)
-    b = np.array(array, copy=False).ravel().view(dt_int)
-    for j in range(0, b.shape[0], 10000):
-        c = b[j:j + 10000]
+    bpc = 1 if dt.kind == 'S' else 4
+    dt_int = "{0}{1}u{2}".format(dt.itemsize // bpc, dt.byteorder, dt.bpc)
+    b = np.array(array, copy=False).view(dt_int)
+    # For optimal speed, work in chunks of the internal ufunc buffer size.
+    bufsize = np.getbufsize()
+    # Attempt to make it a 1-D array so that the chunks have known size.
+    # We cannot use ravel, since that will copy for non-contiguous arrays.
+    # and we need a view of the input array. The code will work on non-1D
+    # arrays, although the chunks will now be larger.
+    if b.ndim > 1:
+        try:
+            b.shape = -1
+        except AttributeError:
+            pass
+    for j in range(0, b.shape[0], bufsize):
+        c = b[j:j + bufsize]
+        # mask which will tell whether we're in a sequence of trailing spaces.
         mask = np.ones(c.shape[:-1], dtype=bool)
+        # loop over the characters in the strings, in reverse order.
         for i in range(-1, -c.shape[-1], -1):
             mask &= c[..., i] == 32
             c[..., i][mask] = 0