--- a/astropy/io/fits/fitsrec.py
+++ b/astropy/io/fits/fitsrec.py
@@ -1307,10 +1307,7 @@ def _rstrip_inplace(array):
     # equal the number of characters in each string.
     bpc = 1 if dt.kind == 'S' else 4
     dt_int = "{0}{1}u{2}".format(dt.itemsize // bpc, dt.byteorder, bpc)
-    if isinstance(array, np.chararray):
-        b = np.array(array, copy=False).view(dt_int)
-    else:
-        b = array.view(dt_int)
+    b = array.view(dt_int, np.ndarray)
     # For optimal speed, work in chunks of the internal ufunc buffer size.
     bufsize = np.getbufsize()
     # Attempt to have the strings as a 1-D array to give the chunk known size.