--- a/astropy/io/fits/hdu/table.py
+++ b/astropy/io/fits/hdu/table.py
@@ -112,10 +112,9 @@ class _TableLikeHDU(_ValidHDU):
             zeros/blanks.
 
         character_as_bytes : bool
-            Whether to represent string columns using byte arrays. By default
-            this is `False` and byte arrays are represented as unicode arrays.
-            Setting this to `True` is much more memory-efficient if unicode
-            arrays are not needed.
+            Whether to return bytes for string columns when accessed from the
+            HDU. By default this is `False` and (unicode) strings are returned,
+            but for large tables this may use up a lot of memory.
 
         Notes
         -----
@@ -127,7 +126,7 @@ class _TableLikeHDU(_ValidHDU):
         coldefs = cls._columns_type(columns)
         data = FITS_rec.from_columns(coldefs, nrows=nrows, fill=fill,
                                      character_as_bytes=character_as_bytes)
-        hdu = cls(data=data, header=header, **kwargs)
+        hdu = cls(data=data, header=header, character_as_bytes=character_as_bytes, **kwargs)
         coldefs._add_listener(hdu)
         return hdu