This is a follow-up to #6821 and speeds up specifically writing tables with string columns.

At the moment, if a table contains a string column, that string column is first decoded by io.fits to a unicode array (if not already the case) before getting re-encoded to a byte array. This means that if the original column was a byte string column (which one could get after reading in a FITS file with Table.read), there is an unnecessary conversion to unicode and back. Furthermore, there was previously a Python loop over all string elements to remove trailing whitespace. This PR uses the ``character_as_bytes`` functionality developed in #6821 to avoid the conversion to unicode in the case a byte string array is passed, and also includes a significantly sped up ``_rstrip_replace`` provided by @mhvk in https://github.com/astropy/astropy/issues/6906.

@mhvk - I feel bad including your code without assigning the commit to you, so if you want the credit, feel free to edit this PR and change the existing commit to be under your user!

Test code:

```python
import numpy as np
from astropy.table import Table

N = 10_000_000

t = Table()
t['floats1'] = np.random.random(N)
t['ints1'] = np.random.randint(0, 100, N)
t['strings'] = b'some strings'

t.write('test_write.fits', overwrite=True)
```

The table initialization takes 1.5s, and the writing of the table itself takes 22.1s before this PR and 2.1s after, giving a speedup of a factor of 10x, and the peak memory usage goes down by a factor of ~2.5x:

![perf](https://user-images.githubusercontent.com/314716/33481186-53b06654-d68b-11e7-98e9-2189c376cb05.png)

(the 'speed up ``_rstrip_replace``' curve also includes ``character_as_bytes``)

If I set ``t['strings']`` to a unicode column instead of bytes column:

```
t['strings'] = 'some strings'
```

Things are also faster with this PR:

![perf_u](https://user-images.githubusercontent.com/314716/33481456-778a4ad0-d68c-11e7-8c63-c1d811747864.png)

(the 'speed up ``_rstrip_replace``' curve also includes ``character_as_bytes``)

In both cases writing a table still seems to double the memory used by the table, but I'm not sure if we can avoid that. One use case I'd like to test is whether if I have a larger-than-memory table I can read it in with memory mapping, modify a cell, and write it out again without using much memory.

Now we might want to consider whether to put some of the optimizations lower down in io.fits - for instance if doing ``BinTableHDU.from_columns`` with a byte array, it will be converted to unicode internally then back to bytes as mentioned above. Check this out:

```python
import numpy as np
from astropy.io.fits import BinTableHDU
x = np.repeat(b'a', 100_000_000)
array = np.array(x, dtype=[('col', 'S1')])
hdu = BinTableHDU.from_columns(array)
print(hdu.dtype)
````

This takes almost a minute and 2.2Gb of memory:

![hdu_false](https://user-images.githubusercontent.com/314716/33481918-8a96fb1c-d68e-11e7-95d9-b97731d4a288.png)

The output for the dtype is ``(numpy.record, [('col', 'S1')])``. If I call it with:

```python
hdu = BinTableHDU.from_columns(array, character_as_bytes=True)
```

it takes 2 seconds and 350Mb of memory:

![hdu_true](https://user-images.githubusercontent.com/314716/33481916-87f89a50-d68e-11e7-8a3a-b9d7b55f749b.png)

And produces the exact same output. So maybe ``BinTableHDU.from_columns`` shouldn't even have a ``character_as_bytes`` option and it should always default to ``True``. But I need to think about the implications about this some more and would value input from FITS experts (e.g. @saimn @MSeifert04).