@mhvk 2017-12-01
@astrofrog - no worry about credit! I did edit your version a little, as I realized `ravel()` could go wrong in some cases: if the array is not contiguous in memory, it will make a copy. So, now I try to do the same by setting the shape, and if it fails just let the array be (i.e., it will use somewhat more memory, but I think one really needs to try to hit that! a proper fits column should never have the problem).