@saimn 2017-12-01
I canceled Travis because it would fail with `dt.bpc`.