@saimn 2017-12-01
Apart from the change in `rstrip_replace` this is really straightforward, just avoiding useless bytes/unicode conversion, wow ! For `rstrip_replace`, the new implementation is less readable for sure, but given the gain we can live with it ;).