@astrofrog 2017-12-01
@mhvk - could you fix ``_rstrip_replace`` directly in this branch? (thanks for the previous commit!)

@saimn - what do you think about my question regarding the following example:

```
import numpy as np
from astropy.io.fits import BinTableHDU
x = np.repeat(b'a', 100_000_000)
array = np.array(x, dtype=[('col', 'S1')])
hdu = BinTableHDU.from_columns(array)
print(hdu.dtype)
```

(see towards the end of the PR description)