@astrofrog 2017-12-01
@mhvk - thanks! I wonder whether we might want to have a centralized place in Astropy to keep Numpy helper functions like this (or in other words functions that try and get around Numpy performance or functionality issues)?