@mhvk 2017-12-01
@astrofrog - ideally we upstream improvements. Though this particular one is so specific to stripping spaces that I think it is fine to keep it in astropy. It may indeed make sense to move it to `utils`...