@astrofrog 2017-12-01
@mhvk - I agree about upstreaming improvements, but sometimes we want the improvement now versus waiting for the fix to be in the oldest supported version of Numpy, so I meant there could be a place where things like that live (and if they never make it upstream, that's fine too)