@astrofrog 2017-12-01
Related to ``chararray``, I found this function:

```
def _get_recarray_field(array, key):
    """
    Compatibility function for using the recarray base class's field method.
    This incorporates the legacy functionality of returning string arrays as
    Numeric-style chararray objects.
    """

    # Numpy >= 1.10.dev recarray no longer returns chararrays for strings
    # This is currently needed for backwards-compatibility and for
    # automatic truncation of trailing whitespace
    field = np.recarray.field(array, key)
    if (field.dtype.char in ('S', 'U') and
            not isinstance(field, chararray.chararray)):
        field = field.view(chararray.chararray)
    return field
```

I wonder whether we might want to consider getting rid of chararrays in io.fits. As indicated in the numpy docs, chararrays are not really recommended anymore and are pretty ancient, so it might be confusing to still return them to users. We also now only support Numpy >= 1.10 which the above comment alludes to. I'm not entirely sure what we would risk breaking in practice for users.