@bsipocz 2017-12-01
> @mhvk - thanks! I wonder whether we might want to have a centralized place in Astropy to keep Numpy helper functions like this (or in other words functions that try and get around Numpy performance or functionality issues)?

There used to be ``numpycompat.py`` in utils/compat, and also there is still a numpy library there.  I wonder whether these can live there? They just need some documentation if migrated upstream about the versions, so we know when it's OK to remove them.