@mhvk 2017-12-01
Yes, `utils.compat.numpy` is close, but at the time I created it we agreed we would only use it for things that were in the process of being upstreamed, i.e., the compatibility was meant to indicate that these things overrode numpy functions.

I think a new `utils.numpy` might be the way forward for little routines such as this. Probably could start with some stuff from `utils.misc` (`NumpyRNGContext`, `ShapedLikeNDArray`, `check_broadcast`, `IncompatibleShapeError`, `dtype_bytes_or_chars`).