@mhvk 2017-12-01
@astropy - I think it would be good to get rid of `chararray` - do raise a separate issue for it!