@bsipocz 2017-12-01
What about having it private, ``utils._numpy``, so even if they may their way upstream, we can just move them around to compat, without the need of having to think about deprecation, etc for downstream users?