@saimn 2017-12-01
@astrofrog - I started a long answer but now I doubt, the array is converted to unicode when you *create* the  `BinTableHDU` with `from_columns` ? I thought that it stores the byte array and convert to unicode only if the column is accessed, but I need to look more in detail.