@saimn 2017-12-01
About chararray deprecation, there is already an issue: #3862