@saimn 2017-12-02
Ok, I understand a bit better now ;), the data is indeed converted when creating the BinTableHDU... 
But the issue with `BinTableHDU.from_columns(array, character_as_bytes=True)` is that you disable completely the bytes to unicode conversion, so if you print the column (or use it) you get bytes.
I think we need to find a way to disable the conversion for inputs (to be able to store bytes internally, as Table do) but still have the conversion when the data is used.