@astrofrog 2017-12-02
@saimn - ok thanks, this is helpful. I think the reason the strings get converted to unicode even if you pass a bytes array in is because of the following blocks of code:

```
                # Make the ndarrays in the Column objects of the ColDefs
                # object of the HDU reference the same ndarray as the HDU's
                # FITS_rec object.
                for idx, col in enumerate(self.columns):
                    col.array = self.data.field(idx)
```

and

```
        # Now replace the original column array references with the new
        # fields
        # This is required to prevent the issue reported in
        # https://github.com/spacetelescope/PyFITS/issues/99
        for idx in range(len(columns)):
            columns._arrays[idx] = data.field(idx)
```

Specifically, doing ``data.field(idx)`` causes ``_convert_other`` to be called (see inside the definition for ``field``), which in turn contains the clause about decoding the array. To be honest I'm a bit lost with all the different references to columns, and including what the purpose of ``columns._arrays`` is. Currently the comments above are actually wrong if the data is converted - that is ``columns._arrays`` and ``col.array`` are NOT references to the underlying data since they might be converted.

One thing that isn't clear to me is whether the following current behavior is buggy (this is before this PR):

```
In [1]: import numpy as np
   ...: from astropy.io.fits import BinTableHDU
   ...: x = np.repeat(b'a', 10_000_000)
   ...: array = np.array(x, dtype=[('col', 'S1')])
   ...: hdu = BinTableHDU.from_columns(array)
   ...: 

In [2]: hdu.data
Out[2]: 
FITS_rec([('a',), ('a',), ('a',), ..., ('a',), ('a',), ('a',)], 
      dtype=(numpy.record, [('col', 'S1')]))

In [3]: hdu.data['col']
Out[3]: 
chararray(['a', 'a', 'a', ..., 'a', 'a', 'a'], 
      dtype='<U1')
```

Specifically I would expect that since I explicitly initialized the BinTableHDU with a byte/string array, there should be no automatic conversion to unicode (in fact the ``Out[2]`` output is confusing as it indicates the dtype is S1 but the output itself is unicode).

At the end of the day, I think you are right: we should never be storing unicode arrays inside the FITS code, and should only ever decode to unicode on output. I do think we need to decide about the behavior of the last example above, that is if we do initialize something with a bytes array explicitly, should it ever be turned into unicode automatically.